refactor: エラーのlog変換処理をloggerに移す
